split logic and rendering for DrinkReminder
