+++ src/components/layout/drink-reminder/index.tsx
@@ -1,47 +1,11 @@
-import { FC, useEffect, useState } from "react";
-import { StorageUtility } from "../../../utilities/storage";
-
+import { FC } from "react";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faDroplet } from "@fortawesome/free-solid-svg-icons";
-
-const storage = new StorageUtility<number>("WATER_TIME");
-
-const ONE_MINUTE = 60_000;
-const ONE_HOUR = 3_600_000;
+import { useLogic } from "./logic";
 
 const DrinkReminder: FC = () => {
 
-   const [remind, setRemind] = useState(false);
-
-   const click = () => {
-      storage.set(Date.now());
-      setRemind(false);
-   };
-
-   const check = () => {
-      const now = Date.now();
-      const lastTime = storage.get();
-
-      if (lastTime) {
-         const diff = now - lastTime;
-         if (diff > ONE_HOUR) {
-            setRemind(true);
-         }
-      } else {
-         storage.set(now - ONE_HOUR);
-         setRemind(true);
-      }
-   };
-
-   useEffect(() => {
-      check();
-
-      const interval = setInterval(check, ONE_MINUTE);
-      return () => {
-         clearInterval(interval);
-      };
-
-   }, []);
+   const { remind, click } = useLogic();
 
    if (!remind) {
       return <></>;
